// apps/api/src/modules/auth/dto/login.dto.ts
import { IsEmail, IsString, MinLength } from 'class-validator';

export class LoginDto {
  @IsEmail({}, { message: 'Please provide a valid email address' })
  email: string;

  @IsString()
  @MinLength(1, { message: 'Password is required' })
  password: string;
}

export class LoginResponseDto {
  redirectPath: string;
  user: {
    id: string;
    fullName: string;
    email: string;
    defaultVertical: string;
  };
}